Update to the latest version of Transcrypt patched with screeps modifications!
Notably, this update changes the default behavior of literal dict construction `{'x': 'y'}` from creating a python JS 'dict' class to just creating a raw JS object.

This means that .keys(), .get(), etc. no longer exist on literals created with `{...}`, and `dict({...})` must be used instead of that is desired.