Modified build.py to accomodate file_expander.py
File expansion can be invoked with the -e or --expand-files switch; by default its off.  A property, source_dir was added to the Configuration class in an effort to modify build.py's source code as little as possible.  The property itself is used only when needed, namely, run_transcrypt() and copy_artifacts().