Revamp documentation and setup!
With the new update removing the `virtualenv` dependency, we have the
potential for a greatly simplified setup (no commands!). It should also
remove the need for anaconda, though that section is left in in case it
works better in other use cases.

This also tries to simplify the README, and point users more directly
to the software setup guide. Maybe help people approach the project?